Added some styles and found a useState bug
Styled task blocks, imported some Google Font
Material Images for resetting timer, editing
the task, and deleting the task.

Found a bug where when the topmost task in TaskList.js is deleted,
the CSS state continues onto the next task and if the 2nd task is
running while the topmost is not, the JS state goes from the topmost
to the 2nd, making it impossible to pause.